tests/submission/W3C/domhtml/HTMLBodyElement03.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Body</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-body-bgcolor" rel="help">
<link rel="author" title="NIST">
</head>
<body bgcolor="#ffff00"> 
<p>Hello, World</p>


<script>
var domtest = async_test("The bgColor IDL attribute of the body element must reflect the element's bgcolor content attribute.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vbgcolor;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("body");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vbgcolor = testNode.bgColor;

      assert_equals(vbgcolor,"#ffff00");
       
    domtest.done();
}));

</script>
<p>The bgColor IDL attribute of the body element must reflect the element's bgcolor content attribute.</p>
<div id="log"></div>
</body></html>
